Rename SelectWithFilterReducers for withFilterReducers
